CCCP-523, feat: implement HookExecutor
