+++ primitives/src/contracts/mod.rs
@@ -2,6 +2,7 @@ pub mod authority;
 pub mod bitcoin_socket;
 pub mod blaze;
 pub mod chainlink_aggregator;
+pub mod erc20;
 pub mod hooks;
 pub mod oracle;
 pub mod registration_pool;
@@ -10,6 +11,7 @@ pub mod relay_queue;
 pub mod relayer_manager;
 pub mod socket;
 pub mod socket_queue;
+pub mod vault;
 
 use alloy::{
 	primitives::{B256, Bytes, FixedBytes, Signature, b256},
